app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
multidict==6.0.4
openai==0.28.0
packaging==23.1
pluggy==1.3.0
pydantic==2.3.0
pydantic-core==2.6.3
pytest==7.4.2
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
